# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Classmate API specification
servers:
  - url: /api
    description: Base API path
security:
  - sessionCookie: []
...
paths:
...
  /courses/{courseId}/enrollments:
    post:
      operationId: enrollStudent
      tags: [enrollments]
      summary: Enroll a student into a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollStudentBody"
      responses:
        "201":
          description: Enrollment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
        "404":
          description: Course or student not found
        "409":
          description: Student already enrolled
        "422":
          description: Course is not active
  /courses/{courseId}/enrollments/{studentId}:
    delete:
      operationId: unenrollStudent
      tags: [enrollments]
      summary: Unenroll a student from a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Enrollment removed
        "404":
          description: Enrollment not found

  # ---- ASSIGNMENTS ----
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List lesson notes/topics
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a lesson note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note/lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  # ---- ANNOUNCEMENTS ----
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session cookie set by the server after login
  schemas:
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        studentId:
          type: integer
        enrolledAt:
          type: string
        enrolledBy:
          type: integer
        isActive:
          type: boolean
        droppedAt:
          type: string
          nullable: true
        droppedBy:
          type: integer
          nullable: true
      required: [id, courseId, studentId, enrolledAt, enrolledBy, isActive]
    EnrollStudentBody:
      type: object
      properties:
        studentId:
          type: integer
      required: [studentId]
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        courseId:
          type: integer
        courseName:
          type: string
        topic:
          type: string
        videoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: integer
          nullable: true
        updatedBy:
          type: integer
          nullable: true
      required: [id, title, content, courseId, courseName, topic, createdAt]
    CreateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        courseId:
          type: integer
        topic:
          type: string
        videoUrl:
          type: string
          nullable: true
      required: [title, content, courseId, topic]
    UpdateNoteBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        topic:
          type: string
        videoUrl:
          type: string
...