# lib/api-zod/src/generated/types/enrollment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Classmate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Enrollment {
  id: number;
  courseId: number;
  studentId: number;
  enrolledAt: string;
  enrolledBy: number;
  isActive: boolean;
  droppedAt?: string | null;
  droppedBy?: number | null;
}